add:gyb-nb-service-gy

+++ gyb-parent/gyb-framework/gyb-model/src/main/java/com/marveliu/framework/model/gy/gy_skill.java
@@ -32,7 +32,7 @@ public class gy_skill extends BaseModel implements Serializable {
     @Column
     @Comment("技能等级")
     @ColDefine(type = ColType.INT)
-    private int skilllevel;
+    private float skilllevel;
 
     // 参照 技能和系统雇员
     @One(field = "skillid")
@@ -40,6 +40,7 @@ public class gy_skill extends BaseModel implements Serializable {
 
     @One(field = "gyid")
     private gy_inf gyinf;
+
 
     public long getId() {
         return id;
@@ -65,11 +66,11 @@ public class gy_skill extends BaseModel implements Serializable {
         this.skillid = skillid;
     }
 
-    public int getSkilllevel() {
+    public float getSkilllevel() {
         return skilllevel;
     }
 
-    public void setSkilllevel(int skilllevel) {
+    public void setSkilllevel(float skilllevel) {
         this.skilllevel = skilllevel;
     }
